requirements.txt
backtester==0.7
matplotlib==3.10.1
numpy==2.2.5  # Pinned version
pandas==2.2.3  # Pinned version
psycopg2_binary==2.9.10
asyncpg==0.29.0
pytest==8.3.5
python_binance==1.0.28  # Pinned version
ta==0.11.0  # Pinned version
